Refactored business layer, added Test & moved Menu item to presentation.

+++ CapitalBusConnectService/src/main/java/eu/sanjin/kurelic/cbc/business/services/CityInfoService.java
@@ -1,11 +1,9 @@
 package eu.sanjin.kurelic.cbc.business.services;
 
 import eu.sanjin.kurelic.cbc.business.viewmodel.city.CityInfoItem;
-import eu.sanjin.kurelic.cbc.business.viewmodel.city.CityInfoItems;
 import eu.sanjin.kurelic.cbc.business.viewmodel.info.InfoItems;
 
 import java.util.Locale;
-import java.util.Optional;
 
 public interface CityInfoService {
 
@@ -15,6 +13,6 @@ public interface CityInfoService {
 
     InfoItems getCityLines(int pageNumber, int limit, Locale language);
 
-    int getNumberOfCityLines();
+    Long getNumberOfCityLines();
 
 }
